refactor: 게임에 관련된 데이터를 다루는 클래스를 위한 인터페이스 도출
HistoryDao 클래스를 인터페이스로 도출
DatabaseHistoryDao 와 FakeHistoryDao 를 HistoryDao 의 구현체로 변경
FakeHistoryDao 의 테스트 코드 수정
SpringChessService, SparkChessService 가 HistoryDao 를 의존하도록 변경
