+++ src/main/java/wooteco/chess/dao/FakeHistoryDao.java
@@ -4,7 +4,7 @@ import wooteco.chess.domain.position.MovingPosition;
 
 import java.util.*;
 
-public class FakeHistoryDao {
+public class FakeHistoryDao implements HistoryDao {
     private final Map<Integer, MovingPosition> fakeHistoryDao;
 
     public FakeHistoryDao() {
@@ -16,25 +16,24 @@ public class FakeHistoryDao {
         fakeHistoryDao.put(4, new MovingPosition("b7", "b5"));
     }
 
-    public Map<Integer, MovingPosition> selectAll() {
-        return fakeHistoryDao;
-    }
-
-    public List<MovingPosition> selectMovingPositions() {
+    @Override
+    public List<MovingPosition> selectAll() {
         return Collections.unmodifiableList(
                 new ArrayList<>(fakeHistoryDao.values())
         );
     }
 
+    @Override
     public void clear() {
         fakeHistoryDao.clear();
     }
 
-    public void insert(String start, String end) {
-        insert(new MovingPosition(start, end));
-    }
-
+    @Override
     public void insert(MovingPosition movingPosition) {
         fakeHistoryDao.put(fakeHistoryDao.size() + 1, movingPosition);
+    }
+
+    public void insert(String start, String end) {
+        insert(new MovingPosition(start, end));
     }
 }
